requirements/runtime.txt
faiss-gpu==1.6.1
future
matplotlib
mmcv-full>=1.4.7
numpy
packaging
pandas
scipy
six
terminaltables
sklearn
seaborn
tensorboard
transformers
timm
tqdm
